+++ src/parser/block/table.ts
@@ -1,6 +1,6 @@
 import { undefined, Array } from 'spica/global';
 import { TableParser } from '../block';
-import { union, sequence, some, block, line, validate, focus, context, creator, surround, open, lazy, fmap } from '../../combinator';
+import { union, sequence, some, block, line, validate, focus, creator, surround, open, lazy, fmap } from '../../combinator';
 import { defrag } from '../util';
 import { inline } from '../inline';
 import { shift, push } from 'spica/array';
@@ -11,12 +11,11 @@ import CellParser = RowParser.CellParser;
 
 export const table: TableParser = lazy(() => block(fmap(validate(
   /^\|[^\n]*(?:\n\|[^\n]*){2,}/,
-  context({ syntax: { inline: { media: false } } },
   sequence([
     row(cell(data), true),
     row(cell(alignment), false),
     some(row(cell(data), true)),
-  ]))),
+  ])),
   rows => {
     const [head, alignment] = shift(rows, 2)[0];
     assert(alignment.children.length > 0);